# docs/swagger.yaml
definitions:
  dto.CreateUserInput:
    properties:
      email:
        example: user@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      password:
        example: password123
        type: string
      status:
        enum:
        - online
        - offline
        - delete
        example: online
        type: string
      type:
        enum:
        - origin
        - oauth2
        example: origin
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
  dto.UserDTO:
    properties:
      create_at:
        example: "2025-07-17 15:20:41"
        type: string
      email:
        example: user@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      is_admin:
        example: true
        type: boolean
      status:
        example: online
        type: string
      type:
        example: origin
        type: string
      u_id:
        example: 123
        type: integer
      update_at:
        example: "2025-07-17 15:20:41"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  models.Group:
    properties:
      createdAt:
        type: string
      description:
        type: string
      gid:
        type: integer
      groupName:
        type: string
      updatedAt:
        type: string
    type: object
  models.Project:
    properties:
      createdAt:
        type: string
      description:
        type: string
      gid:
        description: 'foreign key: group_list.g_id'
        type: integer
      pid:
        type: integer
      projectName:
        type: string
      updatedAt:
        type: string
    type: object
  models.UserWithSuperAdmin:
    properties:
      create_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      is_super_admin:
        type: boolean
      status:
        type: string
      type:
        type: string
      u_id:
        type: integer
      update_at:
        type: string
      username:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.MessageResponse:
    properties:
      message:
        type: string
    type: object
  response.TokenResponse:
    properties:
      token:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /audit/logs:
    get:
      description: Retrieve audit logs filtered by optional parameters like user_id,
        resource_type, action, time range, with pagination support.
      parameters:
      - description: User ID to filter logs by user
        example: 123
        in: query
        name: user_id
        type: integer
      - description: Resource type to filter
        example: '"pod"'
        in: query
        name: resource_type
        type: string
      - description: Action type to filter
        example: '"create"'
        in: query
        name: action
        type: string
      - description: Start time in RFC3339 format, e.g. 2023-01-01T00:00:00Z
        example: '"2023-01-01T00:00:00Z"'
        in: query
        name: start_time
        type: string
      - description: End time in RFC3339 format, e.g. 2023-02-01T00:00:00Z
        example: '"2023-02-01T00:00:00Z"'
        in: query
        name: end_time
        type: string
      - description: Max number of records to return (default 100, max 1000)
        example: 100
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        example: 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of audit logs
          schema:
            items:
              type: object
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Query audit logs
      tags:
      - audit
  /groups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all groups
      tags:
      - groups
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Group name
        in: formData
        name: group_name
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden (reserved name)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Group deleted
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: Invalid group id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden to delete 'super' group
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete group by ID
      tags:
      - groups
    get:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Invalid group id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group by ID
      tags:
      - groups
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group name
        in: formData
        name: group_name
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update group by ID
      tags:
      - groups
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JWT token and user info
          schema:
            $ref: '#/definitions/response.TokenResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to generate token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: User login
      tags:
      - auth
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all projects
      tags:
      - projects
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Project name
        in: formData
        name: project_name
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Group ID
        in: formData
        name: g_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - projects
  /projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: Invalid project id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project by ID
      tags:
      - projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Invalid project id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project name
        in: formData
        name: project_name
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Group ID
        in: formData
        name: g_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project by ID
      tags:
      - projects
  /register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: User registration info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Username already taken
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create user
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: User registration
      tags:
      - auth
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserWithSuperAdmin'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid user id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserWithSuperAdmin'
        "400":
          description: Invalid user id
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - multipart/form-data
      description: Partially update user's email, full name, type, status, or password.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Old password (required if updating password)
        in: formData
        name: old_password
        type: string
      - description: New password
        in: formData
        name: password
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Full name
        in: formData
        name: full_name
        type: string
      - description: 'User type: origin or oauth2'
        in: formData
        name: type
        type: string
      - description: 'User status: online, offline, delete'
        in: formData
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated user info
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"